refactor: move reporting API out of core
