+++ src/api/Cargo.toml
@@ -25,6 +25,7 @@ enterprise = [
     "search_service/enterprise",
     "enrichment-data/enterprise",
     "stream/enterprise",
+    "usage_reporting/enterprise",
 ]
 cloud = [
     "enterprise",
@@ -73,6 +74,7 @@ openobserve-api-management.workspace = true
 openobserve-api-query.workspace = true
 openobserve-api-common.workspace = true
 openobserve-mcp.workspace = true
+usage_reporting.workspace = true
 axum.workspace = true
 axum-extra.workspace = true
 tower.workspace = true
