+++ src/api_management/Cargo.toml
@@ -16,6 +16,7 @@ enterprise = [
     "dep:o2_openfga",
     "openobserve-core/enterprise",
     "openobserve-org-storage/enterprise",
+    "usage_reporting/enterprise",
 ]
 cloud = ["enterprise", "o2_enterprise/cloud", "openobserve-core/cloud"]
 
@@ -31,6 +32,7 @@ vrl.workspace = true
 openobserve-core.workspace = true
 openobserve-node.workspace = true
 openobserve-org-storage.workspace = true
+usage_reporting.workspace = true
 anyhow.workspace = true
 axum.workspace = true
 axum-extra.workspace = true
