+++ src/jobs/Cargo.toml
@@ -22,6 +22,7 @@ enterprise = [
     "search_service/enterprise",
     "stream/enterprise",
     "tantivy_utils/enterprise",
+    "usage_reporting/enterprise",
 ]
 cloud = [
     "enterprise",
@@ -53,6 +54,7 @@ promql-service.workspace = true
 search.workspace = true
 search_service.workspace = true
 tantivy_utils.workspace = true
+usage_reporting.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
 arrow.workspace = true
